Fix SQLite autoincrement primary key for cloud deployment and enhance recon error reporting

+++ fastapi-backend/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, Integer, String, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from database import Base
@@ -6,7 +6,7 @@ from database import Base
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     username = Column(String(255), unique=True, nullable=False, index=True)
     email = Column(String(255), unique=True, nullable=False)
     password = Column(String(255), nullable=False)
@@ -15,7 +15,7 @@ class User(Base):
 class Asset(Base):
     __tablename__ = "assets"
 
-    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     hostname = Column(String(255), nullable=True)
     ip_address = Column(String(255), nullable=False, index=True)
     mac_address = Column(String(255), nullable=True)
@@ -27,20 +27,20 @@ class Asset(Base):
 class Port(Base):
     __tablename__ = "ports"
 
-    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     port_number = Column(Integer, nullable=False)
     protocol = Column(String(255), nullable=False, default="tcp")
     service = Column(String(255), nullable=True)
     state = Column(String(255), nullable=True)
-    asset_id = Column(BigInteger, ForeignKey("assets.id"), nullable=False)
+    asset_id = Column(Integer, ForeignKey("assets.id"), nullable=False)
 
     asset = relationship("Asset", back_populates="ports")
 
 class ScanHistory(Base):
     __tablename__ = "scan_history"
 
-    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
-    asset_id = Column(BigInteger, nullable=False)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
+    asset_id = Column(Integer, nullable=False)
     hostname = Column(String(255), nullable=True)
     ip_address = Column(String(255), nullable=False)
     open_port_count = Column(Integer, default=0)
